29 присвоение правильному экземпляру класса исправление getWeight на getStatus  в циклах while изменение типа double на  int в методах feed и drink.

+++ 03_ObjectsAndClasses/Cat/src/Cat.java
@@ -22,12 +22,12 @@ public class Cat
         System.out.println("Meow");
     }
 
-    public void feed(Double amount)
+    public void feed(int amount)
     {
         weight = weight + amount;
     }
 
-    public void drink(Double amount)
+    public void drink(int amount)
     {
         weight = weight + amount;
     }
